+++ test/imu_preint_test.cpp
@@ -19,7 +19,9 @@ TEST(Imu, exprJacobians) {
 }
 
 Eigen::Matrix3d expMap(const Eigen::Vector3d &phi) {
-    return evalImpl(wave::internal::expr<ExpMap>{}, wave::RelativeRotationd{phi}).value();
+    const auto &q =
+      evalImpl(wave::internal::expr<ExpMap>{}, wave::RelativeRotationd{phi});
+    return q.value().matrix();
 }
 
 Eigen::Vector3d logMap(const Eigen::Matrix3d &C) {